Parser test

+++ Code/M1_PrepareTrainData.py
@@ -72,7 +72,7 @@ if Track=='FEDRA':
  data[PM.MC_Event_ID] = data[PM.MC_Event_ID].astype(str)
  data[PM.FEDRA_Track_ID] = data[PM.FEDRA_Track_ID].astype(int)
  data[PM.FEDRA_Track_ID] = data[PM.FEDRA_Track_ID].astype(str)
- data[PM.FEDRA_Track_QUADRANT] = data[PM.FEDRA_Track_QUADRANT].astype(int)
+ #data[PM.FEDRA_Track_QUADRANT] = data[PM.FEDRA_Track_QUADRANT].astype(int)
  data[PM.FEDRA_Track_QUADRANT] = data[PM.FEDRA_Track_QUADRANT].astype(str)
  data['Track_ID'] = data[PM.FEDRA_Track_QUADRANT] + '-' + data[PM.FEDRA_Track_ID]
  data['Mother_ID'] = data[PM.MC_Event_ID] + '-' + data[PM.MC_VX_ID]
